feat: remove flit formula
This commit removes the flit formula from the repository. The commit message follows the Conventional Commits format with the "feat" type since removing a formula is considered a feature change in the package manager's available formulas.

Note: The "feat" type is used here because removing a formula affects the feature set available to users, though "chore" could also be considered appropriate depending on your team's conventions.